implemented delete tag service tests

+++ src/services/Tag/DeleteTagService.ts
@@ -1,4 +1,5 @@
 import { ITagRepository } from '../../repositories/Tag/TagsRepository.interface';
+import { BadRequestError, NotFoundError } from '../../utils/CustomErrors';
 
 class DeleteTagService {
 
@@ -7,8 +8,10 @@ class DeleteTagService {
   ){}
 
   async execute(id: string) {
+    if (!id) throw new BadRequestError('Empty tag id value');
     const deletedTag = await this.tagsRepository.delete(id);
 
+    if (!deletedTag) throw new NotFoundError('Inexistent tag');
     return deletedTag;
   }
 }
